Finished with flask_jwt_extended

+++ flaskSQLALch/code/resources/item.py
@@ -1,7 +1,8 @@
 
 import sqlite3
 
-from flask_jwt import jwt_required
+#https://flask-jwt-extended.readthedocs.io/en/stable/add_custom_data_claims/
+from flask_jwt_extended import jwt_required, get_jwt
 from flask_restful import Resource, reqparse
 from models.item import ItemModel
 
@@ -12,9 +13,14 @@ class Item(Resource):
         required=True,
         help="this field can't be blank"
     )
+    parser.add_argument('store_id',
+        type=int,
+        required=True,
+        help="Every item needs a store id"
+    )
 
     # @jwt_required()
-    def get(self,name):
+    def get(self, name):
 
         # connection = sqlite3.connect("data.db")
         # cursor = connection.cursor()
@@ -32,7 +38,7 @@ class Item(Resource):
         return {"message": "Item not found"}, 404
 
 
-    # @jwt_required()
+    @jwt_required(fresh=True) # needs to be a newly created jwt access_token
     def post(self, name):
 
         # if next(filter(lambda x: x["name"] == name, items), None) is not None:
@@ -45,15 +51,17 @@ class Item(Resource):
         #data = request.get_json() # Don't need the content type header when force=True is on. silent=True returns none if it's not json
 
         # item = {"name": name, "price": data['price']}
-        item = ItemModel(name, data['price'])
+        item = ItemModel(name, data['price'], data['store_id'])
         try:
             # item.insert()
             item.save_to_db()
-        except:
+        except Exception as e:
+            print(data)
+            print(e)
             return {"message": "An error occurred inserting an item"}, 500
 
         # items.append(item)
-        return item, 201
+        return item.json(), 201
 
     @jwt_required()
     def delete(self, name):
@@ -69,13 +77,17 @@ class Item(Resource):
         # connection.close()
 
         # return {'message': "item deleted"}
+        claims = get_jwt()
+
+        if not claims["is_admin"]:
+            return {"message": "Admin privilege required."}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
         return {'message': "item deleted"}
 
 
-    # @jwt_required()
+    # @jwt_required
     def put(self, name):
         # parser = reqparse.RequestParser()
         # parser.add_argument('price',
@@ -95,30 +107,42 @@ class Item(Resource):
 
         if item is None:
             try:
-                updated_item.insert()
+                item = ItemModel(name, data["price"])
 
                 # ItemModel.insert(updated_item)
             except:
                 return {"message": "An error occurred inserting the item"}, 500
         else:
             try:
-                updated_item.update()
+                item.price = data['price']
             except:
                 return {"message": "An error occurred updating the item"}, 500
+        item.save_to_db()
 
-        return updated_item.json()
+        return item.json()
 
 
 class ItemList(Resource):
+    @jwt_required(optional=True)
     def get(self):
-        connection = sqlite3.connect("data.db")
-        connection.cursor()
+        # connection = sqlite3.connect("data.db")
+        # connection.cursor()
 
-        query = "SELECT * from items"
-        result = connection.execute(query)
-        items = []
-        for row in result:
-            items.append({"name": row[0], "price": row[1]})
-        connection.close()
+        # query = "SELECT * from items"
+        # result = connection.execute(query)
+        # items = []
+        # for row in result:
+        #     items.append({"name": row[0], "price": row[1]})
+        # connection.close()
 
-        return {"items": items}
+        # return {"items": items}
+        # return {'items': [item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt()
+        items = list(map(lambda item: item.json(), ItemModel.find_all()))
+
+        if user_id:
+            return {'items': items},200
+        return {
+            "items": [item['name'] for item in items],
+            "message": "more data available if you are logged in and pass your token. "
+        }, 200
